Renamed more problematic namespaces
