+++ ApiDatastructures/Users/GetUser/UsersGetResponseBody.cs
@@ -1,4 +1,4 @@
-namespace AnonKeyBackend.ApiDatastructures.Users.Get;
+namespace AnonKeyBackend.ApiDatastructures.Users.GetUser;
 
 /// <summary>
 /// Response body of a user get request.
